`cancel` -> `terminate`
Instead of tying cancellation to result, replace with
"early termination", which simply propagates a final value
(that may be a result).

This means that we no longer need the `infallible` method.
We also add various postfix methods for processing results
and cancelling on the `Err` value, like `or_cancel`...

```rust
let ok_value = some_result.or_cancel(scope);
```

...and `or_cancel` on spawned tasks, which replaces this with
a task that cancels the scope if it returns `Result`:

```rust
let ok_value = spawned_task.or_cancel(scope);
```

It'd be nice to implement the previous method on any future,
but we would have to know that `Result: !Future`.